+++ internal/tray/notify_other.go
@@ -3,3 +3,7 @@
 package tray
 
 func notifyNative(Notification) error { return nil }
+
+func notificationDiagnostics() []string {
+	return []string{"notifications: not implemented on this platform"}
+}